requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.21.0
scikit-learn>=1.0.0
